+++ CHANGELOG.md
@@ -18,6 +18,7 @@
     * Add `POST /admin/v2/task-group/{task_group_id}/{deactivate|reactivate}` endpoints (\#2062).
     * Fix bug `_get_collection_task_group_activity_status_message` (\#2047).
     * Remove `valutc` validator for timestamps from API schemas, since it does not match with `psycopg3` behavior (\#2064).
+    * Remove `_convert_to_db_timestamp` and add `_raise_if_naive_datetime`: API only accepts timezone aware datetimes as query parameters (\#2068).
 * Database
     * Drop support for sqlite, and remove the `DB_ENGINE` and `SQLITE_PATH` configuration variables (\#2052).
     * Add `TaskGroupActivityV2` table (\#2005).